+++ test/td_df_lib/parser_test.exs
@@ -346,6 +346,111 @@ defmodule TdDfLib.ParserTest do
       assert fields["input_integer"]["value"] == {:error, :invalid_format}
     end
 
+    test "formats values of type dynamic table for bulk upload" do
+      schema = [
+        %{
+          "name" => "Table Field",
+          "label" => "Table Field",
+          "type" => "dynamic_table",
+          "cardinality" => "*",
+          "values" => %{
+            "table_columns" => [
+              %{
+                "cardinality" => "1",
+                "default" => %{"value" => "", "origin" => "user"},
+                "label" => "First Column",
+                "name" => "First Column",
+                "type" => "string",
+                "values" => nil,
+                "widget" => "dropdown"
+              },
+              %{
+                "cardinality" => "?",
+                "default" => %{"origin" => "default", "value" => ""},
+                "label" => "Second Column",
+                "name" => "Second Column",
+                "type" => "string",
+                "values" => nil,
+                "widget" => "string"
+              }
+            ]
+          }
+        }
+      ]
+
+      # Valid table content
+      content = %{
+        "Table Field" => %{
+          "origin" => "file",
+          "value" =>
+            "First Column;Second Column\r\nFirst Field;Second Field\r\nThird Field;Fourth Field"
+        }
+      }
+
+      assert Parser.format_content(%{
+               content: content,
+               content_schema: schema,
+               domain_ids: []
+             }) == %{
+               "Table Field" => %{
+                 "origin" => "file",
+                 "value" => [
+                   %{
+                     "First Column" => %{"value" => "First Field", "origin" => "file"},
+                     "Second Column" => %{"value" => "Second Field", "origin" => "file"}
+                   },
+                   %{
+                     "First Column" => %{"value" => "Third Field", "origin" => "file"},
+                     "Second Column" => %{"value" => "Fourth Field", "origin" => "file"}
+                   }
+                 ]
+               }
+             }
+
+      # Empty table content
+      content = %{
+        "Table Field" => %{
+          "origin" => "file",
+          "value" => ""
+        }
+      }
+
+      assert Parser.format_content(%{
+               content: content,
+               content_schema: schema,
+               domain_ids: []
+             }) == %{"Table Field" => %{"origin" => "file", "value" => []}}
+
+      # Empty values in every row
+      content = %{
+        "Table Field" => %{
+          "origin" => "file",
+          "value" => "First Column;Second Column\r\n;;;;\r\n;Second Value\r\nThird Value"
+        }
+      }
+
+      assert Parser.format_content(%{
+               content: content,
+               content_schema: schema,
+               domain_ids: []
+             }) == %{
+               "Table Field" => %{
+                 "origin" => "file",
+                 "value" => [
+                   %{
+                     "First Column" => %{"value" => "", "origin" => "file"},
+                     "Second Column" => %{"value" => "", "origin" => "file"}
+                   },
+                   %{
+                     "First Column" => %{"value" => "", "origin" => "file"},
+                     "Second Column" => %{"value" => "Second Value", "origin" => "file"}
+                   },
+                   %{"First Column" => %{"value" => "Third Value", "origin" => "file"}}
+                 ]
+               }
+             }
+    end
+
     test "formats values of type table for bulk upload" do
       schema = [
         %{
@@ -744,6 +849,75 @@ defmodule TdDfLib.ParserTest do
              ]
     end
 
+    test "parses dynamic table fields to binary for excel download" do
+      content = %{
+        "table_name" => %{
+          "origin" => "file",
+          "value" => [
+            %{
+              "col1" => %{"value" => "First Field"},
+              "col4" => %{"value" => ["Fourth Field", "Fith Field"]}
+            },
+            %{"col2" => %{"value" => "Second Field"}, "col3" => %{"value" => "Third Field"}}
+          ]
+        }
+      }
+
+      fields = [
+        %{
+          "type" => "dynamic_table",
+          "name" => "table_name",
+          "values" => %{
+            "table_columns" => [
+              %{
+                "cardinality" => "1",
+                "default" => %{"value" => "", "origin" => "user"},
+                "label" => "col1",
+                "name" => "col1",
+                "type" => "string",
+                "values" => nil,
+                "widget" => "string"
+              },
+              %{
+                "cardinality" => "1",
+                "default" => %{"origin" => "default", "value" => ""},
+                "label" => "col2",
+                "name" => "col2",
+                "type" => "string",
+                "values" => nil,
+                "widget" => "string"
+              },
+              %{
+                "cardinality" => "1",
+                "default" => %{"origin" => "default", "value" => ""},
+                "label" => "col3",
+                "name" => "col3",
+                "type" => "string",
+                "values" => nil,
+                "widget" => "string"
+              },
+              %{
+                "cardinality" => "+",
+                "default" => %{"origin" => "default", "value" => [""]},
+                "label" => "col4",
+                "name" => "col4",
+                "type" => "string",
+                "values" => nil,
+                "widget" => "string"
+              }
+            ]
+          }
+        }
+      ]
+
+      assert Parser.append_parsed_fields([], fields, content, xlsx: true) == [
+               [
+                 "col1;col2;col3;col4\nFirst Field;;;Fourth Field|Fith Field\n;Second Field;Third Field;",
+                 {:align_vertical, :top}
+               ]
+             ]
+    end
+
     test "parses table fields to binary for non excel download" do
       content = %{
         "table_name" => %{
@@ -774,6 +948,60 @@ defmodule TdDfLib.ParserTest do
              ]
     end
 
+    test "parses dynamic table fields to binary for non excel download" do
+      content = %{
+        "table_name" => %{
+          "origin" => "file",
+          "value" => [
+            %{"col1" => %{"value" => "First Field"}},
+            %{"col2" => %{"value" => "Second Field"}, "col3" => %{"value" => "Third Field"}}
+          ]
+        }
+      }
+
+      fields = [
+        %{
+          "type" => "dynamic_table",
+          "name" => "table_name",
+          "values" => %{
+            "table_columns" => [
+              %{
+                "cardinality" => "1",
+                "default" => %{"value" => "", "origin" => "user"},
+                "label" => "col1",
+                "name" => "col1",
+                "type" => "string",
+                "values" => nil,
+                "widget" => "string"
+              },
+              %{
+                "cardinality" => "1",
+                "default" => %{"origin" => "default", "value" => ""},
+                "label" => "col2",
+                "name" => "col2",
+                "type" => "string",
+                "values" => nil,
+                "widget" => "string"
+              },
+              %{
+                "cardinality" => "1",
+                "default" => %{"origin" => "default", "value" => ""},
+                "label" => "col3",
+                "name" => "col3",
+                "type" => "string",
+                "values" => nil,
+                "widget" => "string"
+              }
+            ]
+          }
+        }
+      ]
+
+      assert Parser.append_parsed_fields([], fields, content) == [
+               "col1;col2;col3\nFirst Field;;\n;Second Field;Third Field"
+             ]
+    end
+
     test "returns empty string if no rows when parses table" do
       content = %{
         "table_name" => %{
@@ -788,9 +1016,53 @@ defmodule TdDfLib.ParserTest do
           "name" => "table_name",
           "values" => %{
             "table_columns" => [
-              %{"name" => "col1", "mandatory" => true},
-              %{"name" => "col2", "mandatory" => true},
-              %{"name" => "col3", "mandatory" => true}
+              %{
+                "cardinality" => "1",
+                "default" => %{"value" => "", "origin" => "user"},
+                "label" => "col1",
+                "name" => "col1",
+                "type" => "string",
+                "values" => nil,
+                "widget" => "string"
+              },
+              %{
+                "cardinality" => "1",
+                "default" => %{"origin" => "default", "value" => ""},
+                "label" => "col2",
+                "name" => "col2",
+                "type" => "string",
+                "values" => nil,
+                "widget" => "string"
+              },
+              %{
+                "cardinality" => "1",
+                "default" => %{"origin" => "default", "value" => ""},
+                "label" => "col3",
+                "name" => "col3",
+                "type" => "string",
+                "values" => nil,
+                "widget" => "string"
+              }
+            ]
+          }
+        }
+      ]
+
+      assert Parser.append_parsed_fields([], fields, content) == [""]
+    end
+
+    test "returns empty string if no rows when parsing dynamic table" do
+      content = %{"table_name" => %{"origin" => "file", "value" => []}}
+
+      fields = [
+        %{
+          "type" => "dynamic_table",
+          "name" => "table_name",
+          "values" => %{
+            "table_columns" => [
+              %{"name" => "col1"},
+              %{"name" => "col2"},
+              %{"name" => "col3"}
             ]
           }
         }